// src/app/api/[locale]/sms/providers/aws-sns-enum.ts
/**
 * AWS SNS Enums
 * Separated from aws-sns.ts to avoid circular dependency with env.ts
 */

export enum AwsSnsAwsRegions {
  US_EAST_1 = "us-east-1",
  US_WEST_1 = "us-west-1",
  US_WEST_2 = "us-west-2",
  EU_WEST_1 = "eu-west-1",
  EU_CENTRAL_1 = "eu-central-1",
  AP_SOUTHEAST_1 = "ap-southeast-1",
  AP_SOUTHEAST_2 = "ap-southeast-2",
  AP_NORTHEAST_1 = "ap-northeast-1",
  AP_NORTHEAST_2 = "ap-northeast-2",
  AP_SOUTH_1 = "ap-south-1",
  AP_EAST_1 = "ap-east-1",
  CA_CENTRAL_1 = "ca-central-1",
  SA_EAST_1 = "sa-east-1",
  AF_SOUTH_1 = "af-south-1",
  EU_SOUTH_1 = "eu-south-1",
  ME_SOUTH_1 = "me-south-1",
  EU_NORTH_1 = "eu-north-1",
}